+++ internal/engine/common/codeblock.go
@@ -2,6 +2,12 @@ package common
 
 import "github.com/Azure/InnovationEngine/internal/parsers"
 
+const (
+	STATUS_SUCCESS = "success"
+	STATUS_FAILURE = "failure"
+	STATUS_PENDING = "pending"
+)
+
 // State for the codeblock in interactive mode. Used to keep track of the
 // state of each codeblock.
 type StatefulCodeBlock struct {
@@ -12,12 +18,20 @@ type StatefulCodeBlock struct {
 	StdOut          string            `json:"stdOut"`
 	StepName        string            `json:"stepName"`
 	StepNumber      int               `json:"stepNumber"`
-	Success         bool              `json:"success"`
+	Status          string            `json:"success"`
 	SimilarityScore float64           `json:"similarityScore"`
 }
 
 // Checks if a codeblock was executed by looking at the
 // output, errors, and if success is true.
 func (s StatefulCodeBlock) WasExecuted() bool {
-	return s.StdOut != "" || s.StdErr != "" || s.Error != nil || s.Success
+	return s.Status != STATUS_PENDING
+}
+
+func (s StatefulCodeBlock) Succeeded() bool {
+	return s.Status == STATUS_SUCCESS
+}
+
+func (s StatefulCodeBlock) Failed() bool {
+	return s.Status == STATUS_FAILURE
 }
